+++ packages/asc-ui/src/components/Switch/index.ts
@@ -1,0 +1,2 @@
+export { default as SwitchStyle } from './SwitchStyle'
+export { default } from './Switch'