+++ packages/asc-ui/src/index.ts
@@ -93,6 +93,8 @@ import Select, {
   SelectWrapperStyle,
   SelectContentWrapperStyle,
 } from './components/Select'
+import Switch, { SwitchStyle } from './components/Switch'
+
 import * as constants from './components/shared/constants'
 import ShowMoreShowLess from './components/ShowMoreShowLess'
 import Spinner from './components/Spinner/Spinner'
@@ -206,6 +208,7 @@ export {
   Spinner,
   TextArea,
   TextField,
+  Switch,
   Theme,
   ThemeProvider,
   CompactThemeProvider,
@@ -263,6 +266,7 @@ export const styles = {
   HeadingStyle,
   CustomHTMLBlockStyle,
   ArticleStyle,
+  SwitchStyle,
   TagStyle,
   TopTaskLinkStyle,
   CheckboxStyle,