+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Prefix the change with one of these keywords:
 - Added: `ssr` and `toggleAt` props to `FooterSection` component
 - Deprecated: `showAt` and `hideAt` props of `FooterSection` component
 - Fixed: unwanted props bleeding through to HTML attributes when using `Icon` and `Spinner`
+- Added: `Switch` component!️
 
 ## [0.30.0]
 
@@ -33,6 +34,7 @@ Prefix the change with one of these keywords:
 
 ## [0.29.3]
 
+- Added: Switch component
 - Fixed: issue with not programmically changing the active / initial tab on `Tabs` component
 - Deprecated: `initialTab` in `Tabs` component. Please use `activeTab` instead.
 - Added: Quote component